requirements.txt
python-periphery
paho-mqtt
